% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_basereq.R
\name{ad_basereq}
\alias{ad_basereq}
\title{Generate a base request string for the AREAdata database}
\usage{
ad_basereq()
}
\value{
Returns a string containing the root address of the AREAdata dataset.
}
\description{
This string is used as the basis for all calls to AREAdata.
It does not contain any tokens or session ids, and thus can be regenerated at any time.
}
\examples{
basereq <- ad_basereq()

}
\author{
Francis Windram
}
\concept{basereq}
