% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_gadm_id.R
\name{assoc_gadm_id}
\alias{assoc_gadm_id}
\title{Associate other data sources with gadm IDs}
\usage{
assoc_gadm_id(df, lonlat_names = c("Longitude", "Latitude"), basereq = NA)
}
\arguments{
\item{df}{the source data to bind gadm IDs to. This \strong{must} contain decimal lonlat data!}

\item{lonlat_names}{a vector containing the column names of the longitude and latitude columns \strong{IN THAT ORDER}!}

\item{basereq}{the url of the AREAdata database (usually generated by \code{\link[=ad_basereq]{ad_basereq()}}). If \code{NA}, uses the default.}
}
\value{
A matrix of the \code{data} with added gadm columns.
}
\description{
Intelligently bind together data with gadm IDs at all scales.
}
\section{Caching}{

This will \strong{ALWAYS} get and cache gid level 2 data sources. These files are about 80MB total, so if you are running on a metered connection do beware of this.
}

\examples{
\dontrun{
vt_ids <- search_vt(c("Aedes", "aegypti"))
vtdf <- get_vt_byid(vt_ids[(length(vt_ids)-20):length(vt_ids)]) \%>\%
  extract_vt_data(cols = c(
    "DatasetID",
    "Latitude",
    "Longitude",
    "Interactor1Genus",
    "Interactor1Species"
    ), returnunique = TRUE)
vtdf <- vtdf \%>\% assoc_gadm_id(lonlat_names = c("Longitude", "Latitude"))
}


}
\author{
Francis Windram
}
