% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_status.R
\name{check_db_status}
\alias{check_db_status}
\title{Check whether databases are currently online}
\usage{
check_db_status()
}
\description{
Attempt to access all presently supported databases and report if they were accessible.
}
\examples{
\dontrun{
  check_db_status()
}

}
\author{
Francis Windram
}
