% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ohvbd_config.R
\name{check_ohvbd_config}
\alias{check_ohvbd_config}
\title{Print current ohvbd configuration variables}
\usage{
check_ohvbd_config(options_list = NULL)
}
\arguments{
\item{options_list}{An (optional) list of variables to search for.}
}
\value{
TRUE if all desired options are set (though not necessarily turned on), else FALSE
}
\description{
Access ohvbd options and configured variables, and print them to the command line.
}
\examples{
  check_ohvbd_config()

}
\author{
Francis Windram
}
