% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_provenance}
\alias{check_provenance}
\title{Check whether an object is from a given database and complete appropriate messaging}
\usage{
check_provenance(
  obj,
  db,
  altfunc = "fetch",
  altfunc_suffix = NULL,
  objtype = "IDs",
  call = rlang::caller_env()
)
}
\arguments{
\item{obj}{The object to check.}

\item{db}{The database which the object should be from.}

\item{altfunc}{The function name stub (e.g. "fetch", used for messaging). If \code{NULL}, suppresses alternate function suggestion.}

\item{altfunc_suffix}{The function name suffix (e.g. "chunked", used for messaging).}

\item{objtype}{The type of object (used for messaging).}

\item{call}{The env from which this was called (defaults to the direct calling environment).}
}
\description{
Check whether an object is from a given database and complete appropriate messaging
}
\keyword{internal}
