% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{clean_ad_cache}
\alias{clean_ad_cache}
\title{Delete all rda files from ohvbd AREAdata cache}
\usage{
clean_ad_cache(cache_location = "user")
}
\arguments{
\item{cache_location}{location within which to remove rda files.}
}
\description{
Delete all rda files from ohvbd AREAdata cache
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
clean_ad_cache()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
