% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{clean_ohvbd_cache}
\alias{clean_ohvbd_cache}
\title{Delete files from ohvbd cache directories}
\usage{
clean_ohvbd_cache(subdir = NULL, path = NULL, dryrun = FALSE, force = FALSE)
}
\arguments{
\item{subdir}{a subdirectory or list of subdirectories to clean.}

\item{path}{location within which to remove rda files. (Defaults to the standard ohvbd cache location).}

\item{dryrun}{if \code{TRUE} list files that would be deleted, but do not remove.}

\item{force}{do not ask for confirmation before cleaning.}
}
\value{
No return value, called for side effects
}
\description{
Delete files from ohvbd cache directories
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
clean_ad_cache()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
