% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coercedate}
\alias{coercedate}
\title{Try to coerce a date even when bits are missing}
\usage{
coercedate(dates, return_iso = FALSE, nulliferror = FALSE)
}
\arguments{
\item{dates}{The date/s to try and coerce}
}
\value{
Coerced dates.
}
\description{
Try to coerce a date even when bits are missing
}
\keyword{internal}
