% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_place_togid}
\alias{convert_place_togid}
\title{Convert a vector of place names to their equivalent at a different gid level}
\usage{
convert_place_togid(places, gid = 0)
}
\arguments{
\item{places}{A vector of places.}
}
\value{
vector of converted place names or gid codes.
}
\description{
Convert a vector of place names to their equivalent at a different gid level
}
\keyword{internal}
