% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress.R
\name{default_progress_style}
\alias{default_progress_style}
\title{Reimplementation of cli's default progress style}
\usage{
default_progress_style()
}
\value{
list of characters for a progress bar
}
\description{
Reimplementation of cli's default progress style
}
\keyword{internal}
