% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{.format_dir_as_tree}
\alias{.format_dir_as_tree}
\title{Format directory as df ready for tree plotting}
\usage{
.format_dir_as_tree(d)
}
\arguments{
\item{d}{directory to format}
}
\value{
data.frame of file sytem nodes and children
}
\description{
Format directory as df ready for tree plotting
}
\keyword{internal}
