% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_vb_req_id}
\alias{.get_vb_req_id}
\title{Extract request ids from httr2 response objects}
\usage{
.get_vb_req_id(r)
}
\arguments{
\item{e}{A response object}
}
\value{
the id requested from the call
}
\description{
Extract request ids from httr2 response objects
}
\keyword{internal}
