% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohvbd-deprecated.R
\name{extract}
\alias{extract}
\title{Extract specified data from a set of responses (Deprecated)}
\usage{
extract(res, ...)
}
\arguments{
\item{res}{An object of type \link{ohvbd.responses} or \link{ohvbd.ad.matrix} generated from \code{fetch()}
and containing data from one of the supported databases.}

\item{...}{Any arguments to be passed to the underlying extractors (see \code{glean_vt()} and \code{glean_ad()} for specific arguments).}
}
\value{
The extracted data, either as an \code{ohvbd.data.frame} or \code{ohvbd.ad.matrix} object.
}
\description{
This is a convenience method that infers and applies the correct extractor for the input.
}
\note{
\code{extract()} is now deprecated and should not be used. Please use \code{\link[=glean]{glean()}} instead.
}
\author{
Francis Windram
}
\concept{deprecated}
