% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vd.R
\name{extract_vd}
\alias{extract_vd}
\title{Parse data from requests to VecDyn}
\usage{
extract_vd(res, cols = NA, returnunique = FALSE)
}
\arguments{
\item{res}{a list of responses from VecDyn.}

\item{cols}{a character vector of columns to extract from the dataset.}

\item{returnunique}{whether to return only the unique rows within each dataset according to the filtered columns.}
}
\value{
A dataframe containing the extracted data.
}
\description{
Extract the data returned by a call to \code{\link[=get_vd]{get_vd()}}, filter columns of interest, and find unique rows if required.
}
\examples{
\dontrun{
get_vd_dataset_byid(247) \%>\%
  extract_vd(cols=c("species",
                         "sample_start_date",
                         "sample_value"),
                  returnunique=TRUE)
}

}
\author{
Francis Windram
}
\concept{vecdyn}
