% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vt.R
\name{extract_vt}
\alias{extract_vt}
\title{Parse data from requests to VecTraits}
\usage{
extract_vt(res, cols = NA, returnunique = FALSE, check_src = TRUE)
}
\arguments{
\item{res}{a list of responses from VecTraits as an \code{ohvbd.responses} object.}

\item{cols}{a character vector of columns to extract from the dataset.}

\item{returnunique}{whether to return only the unique rows within each dataset according to the filtered columns.}

\item{check_src}{toggle pre-checking of source data.}
}
\value{
An \code{ohvbd.data.frame} containing the requested data.
}
\description{
Extract the data returned by a call to \code{\link[=fetch_vt]{fetch_vt()}}, filter columns of interest, and find unique rows if required.
}
\examples{
\dontrun{
fetch_vt(54) \%>\%
  extract_vt(cols=c("DatasetID",
                         "Interactor1Genus",
                         "Interactor1Species"),
                  returnunique=TRUE)
}

}
\author{
Francis Windram
}
\concept{vectraits}
