% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{fetch}
\alias{fetch}
\title{Fetch specified data from a set of ids}
\usage{
fetch(ids, ...)
}
\arguments{
\item{ids}{An object of type \code{ohvbd.ids} (generated from a search, manually packaged using \code{\link[=ohvbd.ids]{ohvbd.ids()}} or generated by another function).}

\item{...}{Any other arguments to be passed to the underlying fetch functions (see \code{\link[=fetch_vt]{fetch_vt()}} and \code{\link[=fetch_vd]{fetch_vd()}} for specific arguments).}
}
\value{
The downloaded data, as an \code{ohvbd.responses} object.
}
\description{
This is a convenience method that infers and applies the correct fetch function for the input ids.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
find_vt_ids() |> fetch()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
\concept{convenience}
