% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{fetch_citations}
\alias{fetch_citations}
\title{Try to find the relevant citations for a dataset}
\usage{
fetch_citations(dataset, ...)
}
\arguments{
\item{dataset}{An object of type \code{ohvbd.data.frame} generated from \code{\link[=glean]{glean()}}
and containing data from one of the supported databases.}

\item{...}{Any arguments to be passed to the underlying funcs.}
}
\value{
The extracted data, either as an \code{ohvbd.data.frame} or \code{ohvbd.ad.matrix} object.
}
\description{
This tries to extract and simplify the citations from a dataset downloaded using \code{ohvbd}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
search_hub("Ixodes", "vt") |>
  fetch() |>
  glean() |>
  fetch_citations()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
\concept{convenience}
