% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_citations.R
\name{fetch_citations_vd}
\alias{fetch_citations_vd}
\title{Retrieve citations for vecdyn data}
\usage{
fetch_citations_vd(
  dataset,
  redownload = TRUE,
  minimise = FALSE,
  collapse_cols = TRUE
)
}
\arguments{
\item{dataset}{The dataset from which you wish to retrieve citations.}

\item{redownload}{Redownload data if citation columns are missing.}

\item{minimise}{Whether to return one row per citation (rather than one per dataset ID).}

\item{collapse_cols}{Whether to remove completely empty columns.}
}
\value{
\code{ohvbd.data.frame} of citation data
}
\description{
Retrieve citations for vecdyn data either directly from the dataset or by
redownloading the appropriate data.
}
\author{
Francis Windram
}
\concept{vecdyn}
