% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_gbif.R
\name{fetch_gbif}
\alias{fetch_gbif}
\title{Fetch GBIF dataset/s by ID}
\usage{
fetch_gbif(ids, filepath = ".")
}
\arguments{
\item{ids}{a string or character vector of ids (preferably in an \code{ohvbd.ids} object) indicating the particular dataset/s to download.}

\item{filepath}{directory to save gbif download files into.}
}
\value{
Stuff
}
\description{
Retrieve GBIF dataset/s specified by their dataset ID.
}
\examples{
\dontrun{
fetch_gbif("dbc4a3ae-680f-44e6-ab25-c70e27b38dbc")
}

}
\author{
Francis Windram
}
\concept{gbif}
