% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_vd_columns.R
\name{find_vd_columns}
\alias{find_vd_columns}
\title{Get current columns in VecDyn datasets}
\usage{
find_vd_columns(full = FALSE, basereq = NA)
}
\arguments{
\item{full}{whether to return all the data about current fields, or else just return the names}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A character vector (or dataframe) of column information.
}
\description{
Get all the current column names in VecDyn, alongside associated data if desired.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
find_vd_columns()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
\concept{vecdyn}
