% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vt_template.R
\name{generate_vt_template}
\alias{generate_vt_template}
\title{Generate a vectraits template from a short set of survey responses}
\usage{
generate_vt_template()
}
\value{
A character vector containing the column headers of the desired vectraits template or NULL.
}
\description{
Given the large number of fields in vectraits it can be hard to
know which of these you need to fill out. This generator asks a series of
questions to determine what columns should be included in one's dataset.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

generate_vt_template()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
\concept{vectraits}
