% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_curl_err}
\alias{get_curl_err}
\title{Extract curl errors from httr2 error objects}
\usage{
get_curl_err(e)
}
\arguments{
\item{e}{An error object}
}
\value{
the curl error message (if present) or NULL
}
\description{
Extract curl errors from httr2 error objects
}
\keyword{internal}
