% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_vd_ids.R
\name{get_current_vd_ids}
\alias{get_current_vd_ids}
\title{Get current IDs in VecDyn}
\usage{
get_current_vd_ids(basereq)
}
\arguments{
\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=generate_vb_basereq]{generate_vb_basereq()}}.}
}
\value{
A numeric vector of VecDyn dataset IDs.
}
\description{
Get all the current dataset IDs in VecDyn, as a numeric vector.
}
\examples{
\dontrun{
generate_vb_basereq() \%>\%
  get_current_vd_ids()
}

}
\author{
Francis Windram
}
