% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_vt_ids.R
\name{get_current_vt_ids}
\alias{get_current_vt_ids}
\title{Get current IDs in VecTraits}
\usage{
get_current_vt_ids(basereq)
}
\arguments{
\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}.}
}
\value{
A numeric vector of VecTraits dataset IDs.
}
\description{
Get all the current dataset IDs in VecTraits, as a numeric vector.
}
\examples{
\dontrun{
req <- vb_basereq()
get_current_vt_ids(req)
}

}
\author{
Francis Windram
}
