% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{get_db}
\alias{get_db}
\title{Get the db associated with an object}
\usage{
get_db(x, ...)
}
\arguments{
\item{x}{An object to test.}

\item{...}{Any arguments to be passed to the underlying functions (unused).}
}
\value{
The database identifier associated with an object (or NULL if missing).
}
\description{
This function retrieves the provenance information expected by \code{ohvbd} from an object. If there is no provenance then this returns \code{NULL}.
}
\examples{
ids <- ohvbd.ids(c(1,2,3), "vd")
get_db(ids)

}
\author{
Francis Windram
}
\concept{convenience}
