% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{get_default_ohvbd_cache}
\alias{get_default_ohvbd_cache}
\title{Get ohvbd cache locations}
\usage{
get_default_ohvbd_cache(subdir = NULL, create = TRUE)
}
\arguments{
\item{subdir}{The subdirectory within the cache to find/create (optional).}

\item{create}{Whether to create the cache location if it does not already exist (defaults to TRUE).}
}
\value{
ohvbd cache path as a string
}
\description{
Get ohvbd cache locations
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
get_default_ohvbd_cache()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
