% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extract_vd_chunked.R
\name{get_extract_vd_chunked}
\alias{get_extract_vd_chunked}
\title{Get and parse multiple VecDyn datasets by ID in chunks}
\usage{
get_extract_vd_chunked(
  ids,
  chunksize = 20,
  cols = NA,
  returnunique = FALSE,
  rate = 5,
  connections = 1,
  basereq = NA
)
}
\arguments{
\item{ids}{a numeric vector of IDs indicating the particular datasets to download.}

\item{chunksize}{an integer defining the size of chunks to retrieve in one iteration.}

\item{cols}{a character vector of columns to extract from the dataset.}

\item{returnunique}{whether to return only the unique rows within each dataset according to the filtered columns.}

\item{rate}{maximum number of calls to the API per second.}

\item{connections}{number of simultaneous connections to the server at once. Maximum 8. \strong{Do not enable unless you really need to} as this hits the server significantly harder than usual.}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A dataframe containing the requested data.
}
\description{
Retrieve and parse VecDyn datasets specified by their dataset IDs in batches.

This is not usually necessary (generally you just need \code{\link[=get_vd]{get_vd()}}) but allows one to release data that is not in use from memory. If you would like more control on extraction or parsing then it is best to wrap \code{\link[=get_vd]{get_vd()}} and \code{\link[=extract_vd]{extract_vd()}} in your own chunker instead.
}
\examples{
\dontrun{
get_extract_vd_chunked(c(423,424,425), chunksize = 2, rate=5)
}

}
\author{
Francis Windram
}
\concept{vecdyn}
