% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gadm_sfs.R
\name{get_gadm_sfs}
\alias{get_gadm_sfs}
\title{Get gadm mapping shapefiles}
\usage{
get_gadm_sfs(
  gid = 0,
  cache_location = "user",
  refresh_cache = FALSE,
  basereq = NA
)
}
\arguments{
\item{gid}{the spatial scale to retrieve (0 = country-level, 1=province-level...).}

\item{cache_location}{path to cache location (defaults to user directory obtained from \code{\link[tools:userdir]{tools::R_user_dir()}}).}

\item{refresh_cache}{force a refresh of the relevant cached data.}

\item{basereq}{the url of the AREAdata database (usually generated by \code{\link[=ad_basereq]{ad_basereq()}}). If \code{NA}, uses the default.}
}
\value{
A SpatVector (from \code{\link[terra:vect]{terra::vect()}}) of the requested shapefile.
}
\description{
Retrieve AREAdata gadm mapping shapefiles specified by spatial scale (GID).
}
\examples{
\dontrun{
get_gadm_sfs(gid=0)
}

}
\author{
Francis Windram
}
\concept{areadata}
