% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vd.R
\name{get_vd}
\alias{get_vd}
\title{Get VecDyn dataset/s by ID}
\usage{
get_vd(ids, rate = 5, connections = 1, check_src = TRUE, basereq = NA)
}
\arguments{
\item{ids}{a numeric ID or numeric vector of ids indicating the particular dataset/s to download.}

\item{rate}{maximum number of calls to the API per second.}

\item{connections}{number of simultaneous connections to the server at once. Maximum 8. \strong{Do not enable unless you really need to} as this hits the server significantly harder than usual.}

\item{check_src}{toggle pre-checking of source data prior to get.}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A list of \link[httr2:response]{httr2 response} objects.
}
\description{
Retrieve VecDyn dataset/s specified by their dataset ID.
}
\examples{
\dontrun{
get_vd(54)

get_vd(c(423,424,425), rate=5)
}

}
\author{
Francis Windram
}
\concept{vecdyn}
