% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vd_byid.R
\name{get_vd_byid}
\alias{get_vd_byid}
\title{Get VecDyn dataset/s by ID}
\usage{
get_vd_byid(basereq, ids, rate = 5)
}
\arguments{
\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}.}

\item{ids}{a numeric ID or numeric vector of ids indicating the particular dataset/s to download.}

\item{rate}{maximum number of calls to the API per second.}
}
\value{
A list of \link[httr2:response]{httr2 response} objects.
}
\description{
Retrieve VecDyn dataset/s specified by their dataset ID.
}
\examples{
\dontrun{
req <- vb_basereq()
get_vd_byid(req, 54)

vb_basereq() \%>\% get_vd_byid(c(423,424,425), rate=5)
}

}
\author{
Francis Windram
}
