% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vt_current_ids.R
\name{get_vt_current_ids}
\alias{get_vt_current_ids}
\title{Get current IDs in VecTraits}
\usage{
get_vt_current_ids(basereq = NA)
}
\arguments{
\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A numeric vector of VecTraits dataset IDs.
}
\description{
Get all the current dataset IDs in VecTraits, as a numeric vector.
}
\examples{
\dontrun{
get_vt_current_ids()
}

}
\author{
Francis Windram
}
\concept{vectraits}
