% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vt_dataset_byid.R
\name{get_vt_dataset_byid}
\alias{get_vt_dataset_byid}
\title{Get VecTraits dataset by ID}
\usage{
get_vt_dataset_byid(id_to_find, basereq)
}
\arguments{
\item{id_to_find}{a numeric ID indicating the particular dataset to download}

\item{basereq}{an httr2 request object, as generated by \code{\link[=generate_vb_basereq]{generate_vb_basereq()}}}
}
\value{
Returns a named list of two parts:

\verb{$status} being the status code associated with the request

\verb{$data} being the parsed JSON data from the response.
}
\description{
Retrieve a single VecTraits dataset specified by its dataset ID.
}
\examples{
\dontrun{
req <- generate_vb_basereq()
get_vt_dataset_byid(54, req)
}

}
\author{
Francis Windram
}
