% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vt_datasets_multiid.R
\name{get_vt_datasets_multiid}
\alias{get_vt_datasets_multiid}
\title{Get and parse multiple VecTraits datasets by ID}
\usage{
get_vt_datasets_multiid(
  ids_to_find,
  basereq,
  cols = NA,
  returnunique = FALSE,
  sleeptime = 0.2
)
}
\arguments{
\item{ids_to_find}{a numeric vector of IDs indicating the particular datasets to download}

\item{basereq}{an httr2 request object, as generated by \code{\link[=generate_vb_basereq]{generate_vb_basereq()}}}

\item{cols}{a character vector of columns to extract from the dataset}

\item{returnunique}{whether to return only the unique rows within each dataset according to the filtered columns}

\item{sleeptime}{the enforced delay between API requests to limit strain on the server}
}
\value{
Returns a list of two named parts:

\verb{$df} = dataframe containing the requested data

\verb{$errs} = vector of all error messages encountered when requesting the data (such as when the dataset id is invalid)
}
\description{
Retrieve and parse VecTraits datasets specified by their dataset IDs.
}
\examples{
\dontrun{
req <- generate_vb_basereq()
get_vt_datasets_multiid(c(54,55,56), req)
}

}
\author{
Francis Windram
}
