% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{has_db}
\alias{has_db}
\title{Test whether an object has provenance information}
\usage{
has_db(x, ...)
}
\arguments{
\item{x}{An object to test.}

\item{...}{Any arguments to be passed to the underlying functions (unused).}
}
\value{
Whether the data has a provenance tag (as a boolean).
}
\description{
This function tests whether an object has the provenance information expected by \code{ohvbd}.
}
\examples{
ids <- ohvbd.ids(c(1,2,3), "vd")
has_db(ids)

}
\author{
Francis Windram
}
\concept{convenience}
