% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{is_cached}
\alias{is_cached}
\title{Check whether an object has been loaded from cache by ohvbd}
\usage{
is_cached(x)
}
\arguments{
\item{x}{The object to check.}
}
\description{
Check whether an object has been loaded from cache by ohvbd
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
is.cached(c(1,2,3))
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
