% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{list_ohvbd_cache}
\alias{list_ohvbd_cache}
\title{List all ohvbd cached files}
\usage{
list_ohvbd_cache(subdir = NULL, path = NULL, treeview = FALSE)
}
\arguments{
\item{subdir}{a subdirectory or list of subdirectories to list.}

\item{path}{location within which to list files. (Defaults to the standard ohvbd cache location).}

\item{treeview}{display the full cache in a tree structure}
}
\description{
List all ohvbd cached files
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
list_ohvbd_cache()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
