% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{ohvbd_db}
\alias{ohvbd_db}
\alias{ohvbd_db<-}
\title{Database provenance}
\usage{
ohvbd_db(x)

ohvbd_db(x) <- value
}
\arguments{
\item{x}{An object.}

\item{value}{The value to set the db to.}
}
\value{
The database identifier associated with an object (or \code{NULL} if missing).
}
\description{
Retrieve or set the provenance information expected by \code{ohvbd}.
}
\examples{
ids <- ohvbd.ids(c(1,2,3), "vd")
ohvbd_db(ids)

ohvbd_db(ids) <- "vt"
ohvbd_db(ids)

}
\seealso{
\link[=ohvbd_attrs]{Internal attributes}
}
\author{
Francis Windram
}
\concept{convenience}
