% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ohvbd_dryrun}
\alias{ohvbd_dryrun}
\title{Option: dry runs of ohvbd searches}
\description{
Set this option to make ohvbd terminate searches before execution and return the request object instead.
}
\note{
This is usually only useful when debugging, testing, or developing \code{ohvbd}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
options(ohvbd_dryrun = TRUE)
search_hub("Ixodes ricinus")

options(ohvbd_dryrun = NULL)  # Unset dryrun
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
