% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-progress.R
\name{progress_style}
\alias{progress_style}
\title{Reimplementation of cli's progress style shim}
\usage{
progress_style(x)
}
\arguments{
\item{x}{a string to find in \code{\link[cli:cli_progress_styles]{cli::cli_progress_styles()}}}
}
\value{
a formatting list from \code{\link[cli:cli_progress_styles]{cli::cli_progress_styles()}}
}
\description{
Reimplementation of cli's progress style shim
}
\keyword{internal}
