% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{read_ad_cache}
\alias{read_ad_cache}
\title{Retrieve and format error message from failed vt calls}
\usage{
read_ad_cache(path, metric, gid, warn = TRUE)
}
\arguments{
\item{path}{cache path}

\item{metric}{metric to retrieve}

\item{gid}{gid to retrieve}

\item{warn}{Whether to warn if a cached file is older than 6 months}
}
\value{
cached data
}
\description{
Retrieve and format error message from failed vt calls
}
\keyword{internal}
