% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{set_default_ohvbd_cache}
\alias{set_default_ohvbd_cache}
\title{Set the default ohvbd cache location}
\usage{
set_default_ohvbd_cache(d = NULL)
}
\arguments{
\item{d}{The directory to set the cache path to (or NULL to use a default location).}
}
\value{
The path of the cache (invisibly)
}
\description{
Set the default ohvbd cache location
}
\note{
To permanently set a path to use, add the following to your \code{.Rprofile} file:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(ohvbd_cache = "path/to/directory")
}\if{html}{\out{</div>}}

Where \code{path/to/directory} is the directory in which you wish to cache ohvbd files.

You can find a good default path by running \code{\link[=set_default_ohvbd_cache]{set_default_ohvbd_cache()}} with no arguments.
}
\examples{
set_default_ohvbd_cache()

}
\author{
Francis Windram
}
\concept{convenience}
