% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{space_collapse}
\alias{space_collapse}
\title{collapse a list of character strings to a JS space-separated single string}
\usage{
space_collapse(v)
}
\arguments{
\item{v}{a vector to format.}
}
\value{
collapsed string.
}
\description{
collapse a list of character strings to a JS space-separated single string
}
\keyword{internal}
