% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spatvect_to_multipolygon}
\alias{spatvect_to_multipolygon}
\title{Encode spatvector as WKT, and convert to multipolygon if needed}
\usage{
spatvect_to_multipolygon(v)
}
\arguments{
\item{v}{The spatvector to format.}
}
\value{
The formatted wkt.
}
\description{
Encode spatvector as WKT, and convert to multipolygon if needed
}
\keyword{internal}
