% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vd_error_body}
\alias{vd_error_body}
\title{Retrieve and format error message from failed vd calls}
\usage{
vd_error_body(resp)
}
\arguments{
\item{resp}{An errored response to format.}
}
\value{
error string.
}
\description{
Retrieve and format error message from failed vd calls
}
\keyword{internal}
