% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vd_extraction_helper}
\alias{vd_extraction_helper}
\title{Extract a single vd response, including consistent data}
\usage{
vd_extraction_helper(resp, cols = NA)
}
\arguments{
\item{resp}{A response to extract from}
}
\value{
dataframe of all relevant data in the response
}
\description{
Extract a single vd response, including consistent data
}
\keyword{internal}
