% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vd_make_req}
\alias{vd_make_req}
\title{Create a query for a given VD id at a given page}
\usage{
vd_make_req(id, pagenum, rate, url = "", useragent = "", unsafe = FALSE)
}
\arguments{
\item{id}{ID of the vecdyn dataser}

\item{pagenum}{page to retrieve}

\item{rate}{rate limit for requests}

\item{url}{the base url for the vectorbyte API}

\item{useragent}{the user agent string used when contacting vectorbyte}

\item{unsafe}{disable ssl verification (should only ever be required on Linux, \strong{do not enable this by default})}
}
\value{
\link[httr2:request]{httr2 request} containing the requisite call to retrieve the data
}
\description{
Create a query for a given VD id at a given page
}
\keyword{internal}
