% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vt_error_body}
\alias{vt_error_body}
\title{Retrieve and format error message from failed vt calls}
\usage{
vt_error_body(resp)
}
\arguments{
\item{resp}{An errored response to format.}
}
\value{
error string.
}
\description{
Retrieve and format error message from failed vt calls
}
\keyword{internal}
