% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wkt_to_multipolygon}
\alias{wkt_to_multipolygon}
\title{wkt_to_multipolygon}
\usage{
wkt_to_multipolygon(v)
}
\arguments{
\item{v}{The WKT to convert into a multipolygon.}
}
\value{
The multipolygon equivalent of wkt.
}
\description{
wkt_to_multipolygon
}
\keyword{internal}
