% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{write_ad_cache}
\alias{write_ad_cache}
\title{write data from AREAdata to cache file}
\usage{
write_ad_cache(
  d,
  metric = NULL,
  gid = NULL,
  path = NULL,
  format = "rda",
  compression_type = "bzip2",
  compression_level = 9
)
}
\arguments{
\item{d}{data to write.}

\item{metric}{metric downloaded (inferred if not provided).}

\item{gid}{gid of data (inferred if not provided).}

\item{path}{cache path.}

\item{format}{format to store data in (currenly unused).}

\item{compression_type}{type of compression to use when caching.}

\item{compression_level}{level of compression to use while caching.}
}
\value{
Path of cached file (invisibly)
}
\description{
write data from AREAdata to cache file
}
\keyword{internal}
