% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_status.R
\name{check_db_status}
\alias{check_db_status}
\title{Check whether databases are currently online}
\usage{
check_db_status()
}
\value{
TRUE if all DB checks pass, else FALSE
}
\description{
Attempt to access all presently supported databases and report if they were accessible.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
  check_db_status()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
