% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohvbd-deprecated.R
\name{clean_ad_cache}
\alias{clean_ad_cache}
\title{Delete all rda files from ohvbd AREAdata cache (Deprecated)}
\usage{
clean_ad_cache(cache_location = NULL)
}
\arguments{
\item{cache_location}{location of the cache.}
}
\value{
No return value, called for side effects
}
\description{
Delete all rda files from ohvbd AREAdata cache (Deprecated)
}
\note{
\code{\link[=clean_ad_cache]{clean_ad_cache()}} is now deprecated and should not be used. Please use \code{\link[=clean_ohvbd_cache]{clean_ohvbd_cache()}} instead.
}
\author{
Francis Windram
}
\concept{deprecated}
