% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{clean_ohvbd_cache}
\alias{clean_ohvbd_cache}
\title{Delete files from ohvbd cache directories}
\usage{
clean_ohvbd_cache(cache_location = NULL, subdir = NULL, dryrun = FALSE)
}
\arguments{
\item{cache_location}{location within which to remove rda files. (Defaults to the standard ohvbd cache location).}

\item{subdir}{a subdirectory or list of subdirectories to clean.}

\item{dryrun}{if \code{TRUE} list files that would be deleted, but do not remove.}
}
\description{
Delete files from ohvbd cache directories
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
clean_ad_cache()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
