% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{extract}
\alias{extract}
\title{Extract specified data from a set of responses}
\usage{
extract(res, ...)
}
\arguments{
\item{res}{An object of type \code{ohvbd.responses} or \code{ohvbd.ad.matrix} generated from \code{fetch()}
and containing data from one of the supported databases.}

\item{...}{Any arguments to be passed to the underlying extractors (see \code{extract_vt()} and \code{extract_ad()} for specific arguments).}
}
\value{
The extracted data, either as an \code{ohvbd.data.frame} or \code{ohvbd.ad.matrix} object.
}
\description{
This is a convenience method that infers and applies the correct extractor for the input
}
\examples{
\dontrun{
get_vt_current_ids() \%>\% fetch() \%>\% extract(cols=c("Interactor1Species"))
get_ad(use_cache=TRUE) \%>\% extract(targetdate="2020-08-04")
}
}
\author{
Francis Windram
}
