% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vd.R
\name{extract_vd}
\alias{extract_vd}
\title{Parse data from requests to VecDyn}
\usage{
extract_vd(res, cols = NA, returnunique = FALSE)
}
\arguments{
\item{res}{a list of responses from VecDyn as an \code{ohvbd.responses} object.}

\item{cols}{a character vector of columns to extract from the dataset.}

\item{returnunique}{whether to return only the unique rows within each dataset according to the filtered columns.}
}
\value{
An \code{ohvbd.data.frame} containing the requested data.
}
\description{
Extract the data returned by a call to \code{\link[=fetch_vd]{fetch_vd()}}, filter columns of interest, and find unique rows if required.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
fetch_vd(247) |>
  extract_vd(cols=c("species",
                    "sample_start_date",
                    "sample_value"),
             returnunique=TRUE)
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
\concept{vecdyn}
