% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vt.R
\name{extract_vt}
\alias{extract_vt}
\title{Parse data from requests to VecTraits}
\usage{
extract_vt(res, cols = NA, returnunique = FALSE)
}
\arguments{
\item{res}{a list of responses from VecTraits as an \code{ohvbd.responses} object.}

\item{cols}{a character vector of columns to extract from the dataset.}

\item{returnunique}{whether to return only the unique rows within each dataset according to the filtered columns.}
}
\value{
An \code{ohvbd.data.frame} containing the requested data.
}
\description{
Extract the data returned by a call to \code{\link[=fetch_vt]{fetch_vt()}}, filter columns of interest, and find unique rows if required.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
fetch_vt(54) |>
  extract_vt(cols=c("DatasetID",
                    "Interactor1Genus",
                    "Interactor1Species"),
             returnunique=TRUE)
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
\concept{vectraits}
