% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{fetch}
\alias{fetch}
\title{Fetch specified data from a set of ids}
\usage{
fetch(
  ids,
  ...,
  rate = 5,
  connections = 1,
  check_src = TRUE,
  db = NULL,
  basereq = NA
)
}
\arguments{
\item{ids}{An object of type \code{ohvbd.ids} (generated from a search, manually packaged using \code{ohvbd.ids()} or generated by another function).}

\item{...}{Any other arguments to be passed to the underlying fetch functions (see \code{fetch_vt()} and \code{fetch_vd()} for specific arguments).}

\item{rate}{maximum number of calls to the API per second.}

\item{connections}{number of simultaneous connections to the server at once. Maximum 8. \strong{Do not enable unless you really need to} as this hits the server significantly harder than usual.}

\item{check_src}{toggle pre-checking of source data prior to get.}

\item{db}{An override for the db inferred by the data within \code{ids}. This is rarely necessary if you've used \code{ohvbd.ids()} to construct the id vector.}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
The extracted data, as an \code{ohvbd.responses} object.
}
\description{
This is a convenience method that infers and applies the correct fetch function for the input ids.
}
\examples{
\dontrun{
find_vt_ids() \%>\% fetch()
}
}
\author{
Francis Windram
}
\concept{convenience}
