% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_citations.R
\name{fetch_citations_vd}
\alias{fetch_citations_vd}
\title{Retrieve citations for vecdyn data}
\usage{
fetch_citations_vd(dataset, redownload = TRUE)
}
\arguments{
\item{dataset}{}
}
\value{
\code{ohvbd.data.frame} of citation data
}
\description{
Retrieve citations for vecdyn data either directly from the dataset or by
redownloading the appropriate data.
}
\author{
Francis Windram
}
\concept{vecdyn}
