% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_citations.R
\name{fetch_citations_vt}
\alias{fetch_citations_vt}
\title{Retrieve citations for vectraits data}
\usage{
fetch_citations_vt(dataset, redownload = TRUE, minimise = FALSE)
}
\arguments{
\item{dataset}{The dataset from which you wish to retrieve citations.}

\item{redownload}{Redownload data if citation columns are missing.}

\item{minimise}{Whether to return one row per citation (rather than one per dataset ID).}
}
\value{
\code{ohvbd.data.frame} of citation data
}
\description{
Retrieve citations for vectraits data either directly from the dataset or by
redownloading the appropriate data.
}
\author{
Francis Windram
}
\concept{vectraits}
