% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vd_counts.R
\name{fetch_vd_counts}
\alias{fetch_vd_counts}
\title{Fetch VecDyn dataset length by ID}
\usage{
fetch_vd_counts(ids, rate = 5, connections = 2, page_size = 50, basereq = NA)
}
\arguments{
\item{ids}{a numeric ID or numeric vector of ids (preferably in an \code{ohvbd.ids} object) indicating the particular dataset/s to download.}

\item{rate}{maximum number of calls to the API per second.}

\item{connections}{number of simultaneous connections to the server at once. Maximum 8. \strong{Do not enable unless you really need to} as this hits the server significantly harder than usual.}

\item{page_size}{the page size returned by VecDyn (default is 50).}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
A dataframe describing the number of rows and number of pages for the set of ids.
}
\description{
Retrieve length of VecDyn dataset/s specified by their dataset ID.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
fetch_vd_counts(54)

fetch_vd_counts(c(423,424,425))
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
\concept{vecdyn}
