% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_vd_meta.R
\name{fetch_vd_meta}
\alias{fetch_vd_meta}
\title{Find vecdyn metadata}
\usage{
fetch_vd_meta(
  ids,
  rate = 5,
  connections = 2,
  pb_name = "data",
  basereq = vb_basereq()
)
}
\arguments{
\item{ids}{a numeric ID or numeric vector of ids (preferably in an \code{ohvbd.ids} object) indicating the particular dataset/s to download.}

\item{rate}{maximum number of calls to the API per second.}

\item{connections}{number of simultaneous connections to the server at once. Maximum 8. \strong{Do not enable unless you really need to} as this hits the server significantly harder than usual.}

\item{pb_name}{the name of the download progress bar}

\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}. If \code{NA}, uses the default request.}
}
\value{
list of \code{httr2::response} object
}
\description{
Find vecdyn metadata
}
\keyword{internal}
