% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hub.R
\name{filter_db}
\alias{filter_db}
\title{Filter hub search results by db}
\usage{
filter_db(ids, db)
}
\arguments{
\item{ids}{an \code{ohvbd.hub.search} search result from \code{search_hub()}}

\item{db}{a database name as a string. One of \code{"vt"}, \code{"vd"}, \code{"gbif"}, \code{"px"}.}
}
\value{
An \code{ohvbd.ids} vector of dataset IDs.
}
\description{
Retrieve the IDs for any datasets matching the given db
}
\examples{
\dontrun{
search_hub("Ixodes ricinus") |> filter_db("vt")

}

}
\author{
Francis Windram
}
\concept{hub}
