% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{force_multipolygon}
\alias{force_multipolygon}
\title{Force a polygon WKT into multipolygon form}
\usage{
force_multipolygon(wkt)
}
\arguments{
\item{wkt}{The WKT to convert into a multipolygon.}
}
\value{
The multipolygon equivalent of wkt.
}
\description{
Force a polygon WKT into multipolygon form
}
\keyword{internal}
