% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ad.R
\name{get_ad}
\alias{get_ad}
\title{Get AREAdata dataset}
\usage{
get_ad(
  metric = "temp",
  gid = 0,
  use_cache = FALSE,
  cache_location = ".adcache",
  refresh_cache = FALSE,
  basereq = NA
)
}
\arguments{
\item{metric}{the metric to retrieve from areadata.}

\item{gid}{the spatial scale to retrieve (0 = country-level, 1=province-level).}

\item{use_cache}{load files from cache if possible, and save them if not present.}

\item{cache_location}{path to cache location (defaults to \code{./.adcache}).}

\item{refresh_cache}{force a refresh of the relevant cached data.}

\item{basereq}{the url of the AREAdata database (usually generated by \code{\link[=ad_basereq]{ad_basereq()}}). If \code{NA}, uses the default.}
}
\value{
A matrix of the requested data (with added attributes for gid and metric).
}
\description{
Retrieve AREAdata dataset/s specified by metric and spatial scale (GID).
}
\section{Valid metrics}{

The following metrics are valid (alternative names are listed in brackets):
\itemize{
\item \code{temp} (\emph{temperature})
\item \code{spechumid} (\emph{specific humidity})
\item \code{relhumid} (\emph{relative humidity})
\item \code{uv} (\emph{ultraviolet})
\item \code{precip} (\emph{precipitation, rainfall})
\item \code{popdens} (\emph{population density, population})
\item \code{forecast} (\emph{future climate, future})
}
}

\examples{
\dontrun{
get_ad(metric="temp", gid=0)
}

}
\author{
Francis Windram
}
\concept{areadata}
