% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_curl_err}
\alias{get_curl_err}
\title{Extract curl errors from httr2 error objects}
\usage{
get_curl_err(e, returnfiller = FALSE)
}
\arguments{
\item{e}{An error object}

\item{returnfiller}{If true return filler message if no curl message is present.}
}
\value{
the curl error message (if present) or NULL (or a filler message)
}
\description{
Extract curl errors from httr2 error objects
}
\keyword{internal}
