% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vt_byid.R
\name{get_vt_byid}
\alias{get_vt_byid}
\title{Get VecTraits dataset/s by ID}
\usage{
get_vt_byid(basereq, ids, rate = 5)
}
\arguments{
\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=generate_vb_basereq]{generate_vb_basereq()}}.}

\item{ids}{a numeric ID or numeric vector of ids indicating the particular dataset/s to download.}

\item{rate}{maximum number of calls to the API per second.}
}
\value{
A list of \link[httr2:response]{httr2 response} objects.
}
\description{
Retrieve VecTraits dataset/s specified by their dataset ID.
}
\examples{
\dontrun{
req <- generate_vb_basereq()
get_vt_byid(req, 54)

generate_vb_basereq() \%>\% get_vt_byid(c(54, 55, 56), rate=5)
}

}
\author{
Francis Windram
}
