% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohvbd-package.R
\docType{package}
\name{ohvbd-package}
\alias{ohvbd}
\alias{ohvbd-package}
\title{ohvbd: One Health VBD Hub}
\description{
Interface with the One Health VBD Hub and related repositories directly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fwimp/ohvbd}
  \item Report bugs at \url{https://github.com/fwimp/ohvbd/issues}
}

}
\author{
\strong{Maintainer}: Francis Windram \email{francis.windram17@imperial.ac.uk} (\href{https://orcid.org/0000-0002-2129-826X}{ORCID})

}
\keyword{internal}
