% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hub.R
\name{search_hub}
\alias{search_hub}
\title{Search vbdhub.org}
\usage{
search_hub(
  query,
  db = c("vt", "vd", "gbif", "px"),
  fromdate = NULL,
  todate = NULL,
  locationpoly = NULL,
  exact = FALSE,
  withoutpublished = TRUE,
  returnlist = FALSE,
  connections = 8
)
}
\arguments{
\item{query}{a search string}

\item{db}{the databases to search}

\item{fromdate}{the date from which to search (ISO format: yyyy-mm-dd)}

\item{todate}{the date up to which to search (ISO format: yyyy-mm-dd)}

\item{locationpoly}{a polygon or set of polygons in \code{terra::SpatVector} or WKT MULTIPOLYGON format within which to search}

\item{exact}{whether to return exact matches only}

\item{withoutpublished}{whether to return results without a publishing date when filtering by date}

\item{returnlist}{return the raw output list rather than a formatted df}

\item{connections}{the number of connections to use to parallelise queries}
}
\value{
an \code{ohvbd.hub.search} dataframe or a list (if \code{returnlist=TRUE}) containing the search results
}
\description{
Retrieve the IDs for any datasets matching the given search parameters
}
\examples{
\dontrun{
search_hub("Ixodes ricinus")

}

}
\author{
Francis Windram
}
