% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_vt.R
\name{search_vt}
\alias{search_vt}
\title{Search VecTraits by keyword}
\usage{
search_vt(basereq, keywords)
}
\arguments{
\item{basereq}{an \link[httr2:request]{httr2 request} object, as generated by \code{\link[=vb_basereq]{vb_basereq()}}.}

\item{keywords}{either a string of search terms separated by spaces, or a vector of keywords.}
}
\value{
A numeric vector of VecTraits dataset IDs.
}
\description{
Retrieve the IDs for any VecTraits datasets matching the given keywords
}
\examples{
\dontrun{
vb_basereq() \%>\%
  search_vt("Aedes aegypti")

vb_basereq() \%>\%
  search_vt(c("Aedes", "aegypti"))

}

}
\author{
Francis Windram
}
