% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_basereq.R
\name{set_ohvbd_compat}
\alias{set_ohvbd_compat}
\title{Set ohvbd compatability mode to TRUE}
\usage{
set_ohvbd_compat(value = TRUE)
}
\arguments{
\item{value}{The boolean value to set ohvbd_compat to.}
}
\description{
Set ohvbd to disable ssl verification for calls to external APIs.
This should not be needed (and not be performed) unless you are running on a linux machine or are otherwise experiencing SSL issues when using the package!

When in interactive mode, checks with you to make sure you want to do this. Does not check when run in a script.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set_ohvbd_compat()
\dontshow{\}) # examplesIf}
}
\author{
Francis Windram
}
